% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BASFOR.R
\docType{package}
\name{BASFOR}
\alias{BASFOR}
\alias{BASFOR-package}
\title{BASFOR}
\details{
This package contain the BASic FORest model, BASFOR. The model is a process-based model for forest biogeochemistry which simulates the C-, N- and water-cycles.

For details on how to use the packgage, got to index, vignettes, and have a look at the Quickstart vignette. If the vignette does not appear, try to build it via devtools::build_vignettes("BASFOR").
}
\author{
David Cameron and Marcel van Oijen \email{dcam@ceh.ac.uk}
}

