% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialise_BASFOR_general.R
\name{initialise}
\alias{initialise}
\title{Initialise variables needed to run BASFOR}
\usage{
initialise()
}
\description{
Initialise variables needed to run BASFOR
}
\details{
This function initilaises some matrices needed to run BASFOR
}

