% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialise_BASFOR_general.R
\name{plot_output}
\alias{plot_output}
\title{Function for plotting BASFOR model output}
\usage{
plot_output(list_output = list(output), vars = init$outputNames[-(1:3)])
}
\arguments{
\item{list_output}{a list of model outputs}

\item{vars}{Vector of characters strings of those output variable names which are to be plotted.}
}
\description{
Function for plotting BASFOR model output
}
\details{
This function plots BASFOR model output
}
\examples{
output <- run_model
plot_output()
}

