% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialise_BASFOR_general.R
\name{run_model}
\alias{run_model}
\title{Run the forest model BASFOR}
\usage{
run_model(ft = FORTYPE, p = params, w = matrix_weather,
  calf = init$calendar_fert, calN = init$calendar_Ndep,
  calpT = init$calendar_prunT, caltT = init$calendar_thinT, n = NDAYS,
  NOUT = init$NOUT)
}
\description{
Run the forest model BASFOR
}
\details{
This is a wrapper function to run BASFOR.
}

