% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialise_BASFOR_general.R
\name{weather_BASFOR}
\alias{weather_BASFOR}
\title{Run in the weather input data for BASFOR}
\usage{
weather_BASFOR(y = year_start, d = doy_start, n = NDAYS, df_weather)
}
\description{
Run in the weather input data for BASFOR
}
\details{
This function inputs the weather data from a data frame into an R matrix (matrix_weather).
}

