% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BASFOR.R
\docType{package}
\name{BASFOR}
\alias{BASFOR}
\alias{BASFOR-package}
\title{BASFOR}
\details{
This package contain the BASic FORest model, BASFOR.
The model is a process-based model for forest biogeochemistry which simulates the C-, N- and water-cycles.
For details of how to use the package look at the Quickstart vignette.
}
\author{
David Cameron \email{dcam@ceh.ac.uk} and Marcel van Oijen \email{mvano@ceh.ac.uk}
}

