% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialise_BASFOR_general.R
\name{plot_output}
\alias{plot_output}
\title{Function for plotting BASFOR model output}
\usage{
plot_output(list_output = list(output), vars = outputNames[-(1:3)],
  leg = paste("Run", 1:length(list_output)), leg_title = "LEGEND",
  nrow_plot = ceiling(sqrt((length(vars) + 1) * 8/11)),
  ncol_plot = ceiling((length(vars) + 1)/nrow_plot), lty = rep(1,
  length(list_output)), lwd = rep(3, length(list_output)))
}
\description{
Function for plotting BASFOR model output
}
\details{
This function plots BASFOR model output
}

